Resolve Compiler Warnings (#64)

+++ Source/Core/Editor/MenuEntries/GUI_Menu_File/GUI_Menu_File.cpp
@@ -63,7 +63,7 @@ void GUI_Menu_File::Draw() {
                 );
         }
         if (ImGui::MenuItem("Save All Scenes")) {
-            for (int i = 0; i < SceneManager_->Scenes_.size(); i++) {
+            for (int i = 0; (long)i < (long)SceneManager_->Scenes_.size(); i++) {
                 SceneWriter_->ProcessScene(
                     SceneManager_->Scenes_[i],
                     SceneManager_->Scenes_[i]->ScenePath
